services/processor/requirements.txt
celery==5.6.0
redis==7.1.0
kafka-python==2.0.2
snowflake-snowpark-python==1.44.0
python-dotenv==1.2.1
